requirements.txt
anyio==4.6.2.post1
cairocffi==1.7.1
CairoSVG==2.7.1
certifi==2024.8.30
cffi==1.17.1
cssselect2==0.7.0
defusedxml==0.7.1
h11==0.14.0
httpcore==1.0.7
httpx==0.27.2
idna==3.10
numpy==2.1.3
opencv-python==4.10.0.84
pillow==11.0.0
pycparser==2.22
python-telegram-bot==21.7
sniffio==1.3.1
tinycss2==1.4.0
webencodings==0.5.1
